Major python-scripts refactor (WIP)
