+++ python-scripts/xbox/__init__.py
@@ -1,0 +1,9 @@
+# Load the core components
+from .memory import *
+
+# Load helper functions
+#FIXME: Move stuff into namespaces
+from .pe import * #from . import pe
+from .aci import * #from . import aci
+from .apu import * #from . import apu
+from . import dsp
