+++ python-scripts/xbox/interfaces/memory_nxdk_rdt.py
@@ -1,0 +1,1 @@
+assert(False) # Not written yet.. :(
